Splitting SlicerOpenVR project
Although VR is the main theme of the project week, there was a big and vague umbrella project for SlicerOpenVR that included extension build/administration things, as well as more specific interaction design. There should be more VR projects that are narrower in scope and more specific. Also, this way only those people will be involved in the specific projects who are interested in them instead of a long list of names in the umbrella project