change calendar range to PW30 date range
It is confusing that it was still showing PW29, in particular since that
had the same venue, and was also at the end of a month. Although there
is some value in seeing „a typical project week schedule“, at this point
in time I think we should focus on planning and filling in the upcoming
schedule.