PW39: Ensure front-matter values are escaped when needed
The purpose of this commit are:

* revisits the generation of the project README to ensure the YAML
  values associated with the front-matter section are properly escaped.

* speed-up GitHub workflow by avoiding having to download the docker image
  associated with the GitHub Action "cuchi/jinja2-action"

* streamline debugging by outputting the content of "template-data.json"

* remove need for generating the "template-data.json" file in the current
  directory. Since no docker image is used for rendering the template, it
  can be used from the runner temp directory.


It is a follow-up of 66e3d34f (BUG: Remove ':' in project title)

To ensure the values are escaped, the jinja2 package is directly used. This
provides more flexibility as it allows to leverage the "to_yaml" and
"regex_replace" filters provided by the "jinja2-ansible-filters" package.

Note that since to_yaml is originally intended to output a complete
document, we need to make sure to strip "\n" or "\n...\n" from the
text as these do not make ensure in the context of a single YAML value.

For references related to the "...", see these links:
* https://stackoverflow.com/questions/56950391/yaml-end-always-dumped-even-if-yaml-explicit-end-false
* https://stackoverflow.com/questions/66542271/what-does-three-dots-in-a-yaml-yml-mean/66542407#66542407
* https://github.com/yaml/pyyaml/issues/379
* https://github.com/ansible/ansible/issues/17095